Refiz a view no modelo FBV, tambêm dei uma padronizada nos templates
