+++ stock/templates/category_delete.html
@@ -3,24 +3,13 @@
     Excluir
 {% endblock title %}
 {% block body %}
-    <div class="flex items-center justify-center min-h-screen bg-gray-100">
-        <div class="bg-white p-8 rounded shadow-md max-w-md w-full">
-            <h2 class="text-xl font-semibold text-gray-800 mb-4">Confirmar exclusão</h2>
-            <p class="text-gray-600 mb-6">
-                Tem certeza que deseja excluir a categoria
-                <strong>{{ category.name }}</strong>? Essa ação não poderá ser desfeita.
-            </p>
-            <form method="post" class="flex justify-between items-center">
-                {% csrf_token %}
-                <a href="{% url 'category_list' %}"
-                   class="inline-block px-4 py-2 text-sm font-medium text-gray-700 bg-gray-200 rounded hover:bg-gray-300">
-                    Cancelar
-                </a>
-                <button type="submit"
-                        class="px-4 py-2 text-sm font-medium text-white bg-red-600 rounded hover:bg-red-700">
-                    Confirmar exclusão
-                </button>
-            </form>
-        </div>
-    </div>
+    <h2>Confirmar exclusão</h2>
+    <p>
+        Tem certeza que deseja excluir a categoria <strong>{{ category.name }}</strong>? Essa ação não poderá ser desfeita.
+    </p>
+    <form method="post">
+        {% csrf_token %}
+        <a href="{% url 'category_list' %}">Cancelar</a>
+        <button type="submit">Confirmar exclusão</button>
+    </form>
 {% endblock body %}
