+++ stock/templates/ingredient_detail.html
@@ -9,5 +9,4 @@
     <p>{{ ingredient.min_qte }}</p>
     <p>{{ ingredient.measure_unit }}</p>
     <p>{{ ingredient.price }}</p>
-    <p>{{ ingredient.active }}</p>
 {% endblock body %}
