+++ stock/templates/ingredient_update.html
@@ -15,23 +15,22 @@
             {% endfor %}
         </select>
         <label for="qte">Quantidade:</label>
-        <input type="number" name="qte" value="{{ ingredient.qte }}" />
+        <input type="number" name="qte" value="{{ ingredient.qte }}" min="0" />
         <label for="min_qte">Quantidade Mínima:</label>
-        <input type="number" name="min_qte" value={{ ingredient.min_qte }}>
+        <input type="number" name="min_qte" value="{{ ingredient.min_qte }}" min="0" />
         <label for="measure_unit">Unidade de Medida:</label>
         <select name="measure_unit">
-            {% for value, label in form.measure_unit.field.choices %}
+            {% for value, label in measure_units %}
                 <option value="{{ value }}"
-                        {% if form.measure_unit.value == value %}selected{% endif %}>{{ label }}</option>
+                        {% if ingredient.measure_unit == value %}selected{% endif %}>{{ label }}</option>
             {% endfor %}
         </select>
         <label for="price">Preço:</label>
-        <input type="number" name="price" step="0.01" value={{ ingredient.price }}>
-        <label for="status">Status:</label>
-        <select name="active">
-            <option value="True" {% if ingredient.active %}selected{% endif %}>Ativo</option>
-            <option value="False" {% if not ingredient.active %}selected{% endif %}>Inativo</option>
-        </select>
+        <input type="number"
+               name="price"
+               step="0.01"
+               value="{{ ingredient.price }}"
+               min="0" />
         <button type="submit">Alterar</button>
     </form>
 {% endblock body %}
