+++ stock/views.py
@@ -1,237 +1,360 @@
 from django.contrib import messages
-from django.http import HttpResponseRedirect
-from django.urls import reverse
-from django.views.generic import CreateView, DeleteView, DetailView, ListView, UpdateView
+from django.contrib.auth.decorators import login_required
+from django.forms import ValidationError
+from django.shortcuts import get_object_or_404, redirect, render
+from django.views.decorators.http import require_http_methods
 
-from core.mixins import AdminRequiredMixin, LoginRequiredMixin
-
-from .models import Category, Ingredient, Product
-from .services import register_product, update_product
+from .models import Category, Ingredient, Product, ProductIngredient
 
 
-class CategoryCreate(LoginRequiredMixin, AdminRequiredMixin, CreateView):
-    model = Category
-    fields = ["name", "description"]
-    template_name = "category_create.html"
+@login_required
+@require_http_methods(["GET", "POST"])
+def category_create(request):
+    if request.method == "GET":
+        return render(request, "category_create.html")
 
-    def get_success_url(self):
-        messages.success(self.request, "Categoria registrada com sucesso!")
-        return reverse("category_list")
+    else:
+        name = request.POST.get("name")
+        description = request.POST.get("description")
 
+        if Category.objects.filter(name__iexact=name):
+            messages.error(request, "A categoria que deseja cadastrar já existe!")
+            return redirect("category_list")
 
-class CategoryList(LoginRequiredMixin, ListView):
-    model = Category
-    template_name = "category_list.html"
-    context_object_name = "categories"
+        category = Category.objects.create(name=name, description=description)
 
+        category.save()
 
-class CategoryDetail(LoginRequiredMixin, AdminRequiredMixin, DetailView):
-    model = Category
-    template_name = "category_detail.html"
-    context_object_name = "category"
-
-
-class CategoryUpdate(LoginRequiredMixin, AdminRequiredMixin, UpdateView):
-    fields = ["name", "description"]
-    model = Category
-    template_name = "category_update.html"
-    context_object_name = "category"
-
-    def get_success_url(self):
-        messages.success(self.request, "Categoria alterada com sucesso!")
-        return reverse("category_list")
+        messages.success(request, "Categoria criada com sucesso!")
+        return redirect("category_list")
 
 
-class CategoryDelete(LoginRequiredMixin, AdminRequiredMixin, DeleteView):
-    model = Category
-    template_name = "category_delete.html"
-
-    def get_success_url(self):
-        messages.success(self.request, "Categoria deletada com sucesso!")
-        return reverse("category_list")
+@login_required
+@require_http_methods(["GET"])
+def category_list(request):
+    context = {"categories": Category.objects.all()}
+    return render(request, "category_list.html", context)
 
 
-class IngredientCreate(LoginRequiredMixin, AdminRequiredMixin, CreateView):
-    model = Ingredient
-    fields = ["name", "category", "qte", "min_qte", "price", "measure_unit"]
-    template_name = "ingredient_create.html"
-    context_object_name = "ingredient"
+@login_required
+@require_http_methods(["GET"])
+def category_detail(request, id):
+    context = {"category": get_object_or_404(Category, id=id)}
 
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["categories"] = Category.objects.all()
-        return context
-
-    def get_success_url(self):
-        messages.success(self.request, "Ingrediente registrado com sucesso!")
-        return reverse("ingredient_list")
+    return render(request, "category_detail.html", context)
 
 
-class IngredientList(LoginRequiredMixin, ListView):
-    model = Ingredient
-    template_name = "ingredient_list.html"
-    context_object_name = "ingredients"
+@login_required
+@require_http_methods(["GET", "POST"])
+def category_update(request, id):
+    category = get_object_or_404(Category, id=id)
+    context = {"category": category}
 
-    def get_queryset(self):
-        queryset = super().get_queryset()
-        value = self.request.GET.get("value")
-        field = self.request.GET.get("field")
+    if request.method == "GET":
+        return render(request, "category_update.html", context)
 
-        # NOTE: adicionar mais condicionais no filtro do qte
-        if value and field:
-            if field == "name":
-                queryset = queryset.filter(name__icontains=value)
-            elif field == "category":
-                queryset = queryset.filter(category__icontains=value)
-            elif field == "qte":
-                queryset = queryset.filter(qte=value)
-            elif field == "min_qte":
-                queryset = queryset.filter(min_qte=value)
-            elif field == "price":
-                queryset = queryset.filter(price=value)
-            elif field == "active":
-                status = {"ativo": True, "inativo": False}
-                queryset = queryset.filter(active=status[value.lower()])
+    else:
+        name = request.POST.get("name")
 
-        return queryset
+        if Category.objects.filter(name__iexact=name).exclude(id=category.id).exists():
+            messages.error(request, "O novo nome que deseja inserir já está associado a uma categoria")
+            return render(request, "category_update.html", context)
+
+        category.name = name
+        category.description = request.POST.get("description")
+        category.save()
+
+        messages.success(request, "Categoria alterada com sucesso!")
+        return redirect("category_list")
 
 
-class IngredientDetail(LoginRequiredMixin, AdminRequiredMixin, DetailView):
-    model = Ingredient
-    template_name = "ingredient_detail.html"
-    context_object_name = "ingredient"
+@login_required
+@require_http_methods(["GET", "POST"])
+def category_delete(request, id):
+    category = get_object_or_404(Category, id=id)
+
+    if request.method == "GET":
+        context = {"category": category}
+        return render(request, "category_delete.html", context)
+
+    else:
+        category.delete()
+
+        messages.success(request, "Categoria deletada com sucesso!")
+        return redirect("category_list")
 
 
-class IngredientUpdate(LoginRequiredMixin, AdminRequiredMixin, UpdateView):
-    model = Ingredient
-    fields = ["name", "category", "qte", "min_qte", "price", "measure_unit", "active"]
-    template_name = "ingredient_update.html"
-    context_object_name = "ingredient"
+@login_required
+@require_http_methods(["GET", "POST"])
+def ingredient_create(request):
+    if request.method == "GET":
+        context = {
+            "categories": Category.objects.all(),
+            "measure_units": Ingredient._meta.get_field("measure_unit").choices,
+        }
+        return render(request, "ingredient_create.html", context)
 
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["categories"] = Category.objects.all()
-        return context
+    else:
+        name = request.POST.get("name")
+        qte = request.POST.get("qte")
+        min_qte = request.POST.get("min_qte")
+        category_id = request.POST.get("category")
+        measure_unit = request.POST.get("measure_unit")
+        price = request.POST.get("price")
 
-    def get_success_url(self):
-        messages.success(self.request, "Ingrediente alterado com sucesso!")
-        return reverse("ingredient_list")
+        category = get_object_or_404(Category, id=category_id)
 
+        if Ingredient.objects.filter(name__iexact=name).exists():
+            messages.error(request, "O ingrediente que deseja cadastrar já existe!")
+            return redirect("ingredient_list")
 
-class IngredientDelete(LoginRequiredMixin, AdminRequiredMixin, DeleteView):
-    model = Ingredient
-    template_name = "ingredient_delete.html"
+        ingredient = Ingredient.objects.create(
+            name=name,
+            category=category,
+            qte=qte,
+            min_qte=min_qte,
+            measure_unit=measure_unit,
+            price=price,
+        )
 
-    def get_success_url(self):
-        messages.success(self.request, "Ingrediente deletado com sucesso!")
-        return reverse("ingredient_list")
+        ingredient.save()
 
-
-class ProductCreate(LoginRequiredMixin, AdminRequiredMixin, CreateView):
-    model = Product
-    fields = ["name", "ingredients", "price"]
-    template_name = "product_create.html"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["ingredients"] = Ingredient.objects.all()
-        return context
-
-    def form_valid(self, form):
-        product_name = form.cleaned_data["name"]
-        price = form.cleaned_data["price"]
-        selected_ids = self.request.POST.getlist("ingredients")
-        post_data = self.request.POST
-
-        product_result = register_product(self.request, product_name, price, selected_ids, post_data)
-
-        if product_result:
-            return HttpResponseRedirect(self.get_success_url())
-        else:
-            return self.form_invalid(form)
-
-    def form_invalid(self, form):
-        if "name" in form.errors:
-            messages.error(self.request, f"Já existe um produto com esse nome!")
-
-        return super().form_invalid(form)
-
-    def get_success_url(self):
-        messages.success(self.request, "Produto criado com sucesso!")
-        return reverse("product_list")
+        messages.success(request, "Ingrediente cadastrado com sucesso!")
+        return redirect("ingredient_list")
 
 
-class ProductList(LoginRequiredMixin, ListView):
-    model = Product
-    fields = ["name", "price"]
-    template_name = "product_list.html"
-    context_object_name = "products"
+@login_required
+@require_http_methods(["GET"])
+def ingredient_list(request):
+    ingredients = Ingredient.objects.all()
+    categories = Category.objects.all()
 
-    def get_queryset(self):
-        queryset = super().get_queryset()
-        value = self.request.GET.get("value")
-        field = self.request.GET.get("field")
+    field = request.GET.get("field")
+    value = request.GET.get("value")
 
-        if value and field:
-            if field == "name":
-                queryset = queryset.filter(name__icontains=value)
-            elif field == "category":
-                queryset = queryset.filter(category__icontains=value)
-            elif field == "price":
-                quryset = queryset.filter(price=value)
+    if field and value:
+        if field == "name":
+            ingredients = ingredients.filter(name__icontains=value)
+        elif field == "category":
+            ingredients = ingredients.filter(category__name__icontains=value)
+        elif field == "qte":
+            ingredients = ingredients.filter(qte=value)
+        elif field == "min_qte":
+            ingredients = ingredients.filter(min_qte=value)
+        elif field == "price":
+            ingredients = ingredients.filter(price=value)
 
-        return queryset
-
-
-class ProductDetail(LoginRequiredMixin, AdminRequiredMixin, DetailView):
-    model = Product
-    fields = ["name", "ingredients", "price"]
-    template_name = "product_detail.html"
-    context_object_name = "product"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        product = self.get_object()
-        context["products_ingredients"] = product.productingredient_set.all()
-        return context
+    context = {
+        "ingredients": ingredients,
+        "categories": categories,
+        "field": field,
+        "value": value,
+    }
+    return render(request, "ingredient_list.html", context)
 
 
-class ProductUpdate(LoginRequiredMixin, AdminRequiredMixin, UpdateView):
-    model = Product
-    fields = ["name", "ingredients", "price"]
-    template_name = "product_update.html"
-    context_object_name = "product"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["ingredients"] = Ingredient.objects.all()
-        product = self.get_object()
-        context["products_ingredients"] = product.productingredient_set.all()
-        return context
-
-    def form_valid(self, form):
-        product_to_update = self.get_object()
-        new_name = form.cleaned_data["name"]
-        new_price = form.cleaned_data["price"]
-        selected_ids = self.request.POST.getlist("ingredients")
-        post_data = self.request.POST
-
-        product_result = update_product(self.request, product_to_update, new_name, new_price, selected_ids, post_data)
-
-        if product_result:
-            return HttpResponseRedirect(self.get_success_url())
-        else:
-            return self.form_invalid(form)
-
-    def get_success_url(self):
-        messages.success(self.request, "Produto alterado com sucesso!")
-        return reverse("product_list")
+@login_required
+@require_http_methods(["GET"])
+def ingredient_detail(request, id):
+    context = {"ingredient": get_object_or_404(Ingredient, id=id)}
+    return render(request, "ingredient_detail.html", context)
 
 
-class ProductDelete(LoginRequiredMixin, AdminRequiredMixin, DeleteView):
-    model = Product
-    template_name = "product_delete.html"
+@login_required
+@require_http_methods(["GET", "POST"])
+def ingredient_update(request, id):
+    ingredient = get_object_or_404(Ingredient, id=id)
+    context = {
+        "ingredient": ingredient,
+        "categories": Category.objects.all(),
+        "measure_units": Ingredient._meta.get_field("measure_unit").choices,
+    }
+    if request.method == "GET":
+        return render(request, "ingredient_update.html", context)
 
-    def get_success_url(self):
-        messages.success(self.request, "Produto deletado com sucesso!")
-        return reverse("product_list")
+    else:
+        name = request.POST.get("name")
+
+        if Ingredient.objects.filter(name__iexact=name).exclude(id=ingredient.id).exists():
+            messages.error(request, "O novo nome que deseja inserir já está associado a um ingrediente")
+            return redirect("ingredient_list")
+
+        ingredient.name = name
+        category_id = request.POST.get("category")
+        ingredient.category = get_object_or_404(Category, id=category_id)
+        ingredient.qte = request.POST.get("qte")
+        ingredient.min_qte = request.POST.get("min_qte")
+        ingredient.measure_unit = request.POST.get("measure_unit")
+        ingredient.price = request.POST.get("price")
+
+        ingredient.save()
+
+        messages.success(request, "Ingrediente alterado com sucesso!")
+        return redirect("ingredient_list")
+
+
+@login_required
+@require_http_methods(["GET", "POST"])
+def ingredient_delete(request, id):
+    ingredient = get_object_or_404(Ingredient, id=id)
+
+    if request.method == "GET":
+        context = {"ingredient": ingredient}
+        return render(request, "ingredient_delete.html", context)
+
+    else:
+        ingredient.delete()
+
+        messages.success(request, "Ingrediente deletado com sucesso!")
+        return redirect("ingredient_list")
+
+
+@login_required
+@require_http_methods(["GET", "POST"])
+def product_create(request):
+    if request.method == "GET":
+        context = {"ingredients": Ingredient.objects.all()}
+        return render(request, "product_create.html", context)
+
+    else:
+        name = request.POST.get("name")
+
+        if Product.objects.filter(name__iexact=name).exists():
+            messages.error(request, "O produto que deseja criar já existe!")
+            return redirect("product_list")
+
+        price = request.POST.get("price")
+        ingredients_ids = request.POST.getlist("ingredients")
+
+        ingredients_to_create = []
+
+        for ingredient_id in ingredients_ids:
+            try:
+                quantity = float(request.POST.get(f"q-{ingredient_id}"))
+            except ValueError:
+                ingredient_name = Ingredient.objects.get(pk=ingredient_id).name
+                messages.error(request, f"Forneça uma quantidade válida para o ingrediente {ingredient_name}!")
+                return redirect("product_list")
+
+            ingredients_to_create.append((int(ingredient_id), quantity))
+
+        product = Product.objects.create(name=name, price=price)
+
+        for ingredient_id, quantity in ingredients_to_create:
+            ProductIngredient.objects.create(
+                product=product,
+                ingredient_id=ingredient_id,
+                quantity=quantity,
+            )
+
+        messages.success(request, "Produto criado com sucesso!")
+        return redirect("product_list")
+
+
+@login_required
+@require_http_methods(["GET"])
+def product_list(request):
+    products = Product.objects.all()
+
+    field = request.GET.get("field")
+    value = request.GET.get("value")
+
+    if field and value:
+        if field == "name":
+            products = products.filter(name__icontains=value)
+        elif field == "price":
+            try:
+                value = float(value)
+            except ValueError:
+                messages.error(request, "Insira um número válido!")
+                return redirect("product_list")
+
+            products = products.filter(price=value)
+
+    context = {
+        "products": products,
+        "field": field,
+        "value": value,
+    }
+
+    return render(request, "product_list.html", context)
+
+
+@login_required
+@require_http_methods(["GET"])
+def product_detail(request, id):
+    product = get_object_or_404(Product, id=id)
+    context = {
+        "product": product,
+        "products_ingredients": product.productingredient_set.all(),
+    }
+    return render(request, "product_detail.html", context)
+
+
+@login_required
+@require_http_methods(["GET", "POST"])
+def product_update(request, id):
+    product = get_object_or_404(Product, id=id)
+
+    if request.method == "GET":
+        context = {
+            "product": product,
+            "ingredients": Ingredient.objects.all(),
+            "product_ingredients": product.productingredient_set.all(),
+        }
+        return render(request, "product_update.html", context)
+
+    else:
+        name = request.POST.get("name")
+
+        if Product.objects.filter(name__iexact=name).exclude(id=product.id).exists():
+            messages.error(request, "O novo nome que deseja inserir já está associado a um produto!")
+            return redirect("product_list")
+
+        price = request.POST.get("price")
+        selected_ids = request.POST.getlist("ingredients")
+
+        product.name = name
+        product.price = price
+        product.save(update_fields=["name", "price"])
+
+        old_ingredients_ids = set(product.productingredient_set.values_list("ingredient_id", flat=True))
+        new_ingredients_ids = set(int(pk) for pk in selected_ids)
+
+        ids_to_remove = old_ingredients_ids - new_ingredients_ids
+
+        if ids_to_remove:
+            ProductIngredient.objects.filter(product=product, ingredient_id__in=ids_to_remove).delete()
+
+        for ingredient_id in new_ingredients_ids:
+            try:
+                quantity = float(request.POST.get(f"q-{ingredient_id}"))
+            except ValueError:
+                ingredient_name = Ingredient.objects.get(pk=ingredient_id).name
+                messages.error(request, f"Forneça uma quantidade válida para o ingrediente {ingredient_name}")
+                return redirect("product_list")
+
+            ProductIngredient.objects.update_or_create(
+                product=product,
+                ingredient_id=ingredient_id,
+                quantity=quantity,
+            )
+
+        messages.success(request, "Produto alterado com sucesso!")
+        return redirect("product_list")
+
+
+@login_required
+@require_http_methods(["GET", "POST"])
+def product_delete(request, id):
+    product = get_object_or_404(Product, id=id)
+
+    if request.method == "GET":
+        context = {"product": product}
+        return render(request, "product_delete.html", context)
+
+    else:
+        product.delete()
+
+        messages.success(request, "Produto deletado com sucesso!")
+        return redirect("product_list")
